removed extending BaseNrtmServerChannelInitializer
